refactor: SoundBlocks style 및 구조 개선
